Rename file fixoldclaims.py to fexoldfunds.py
